requirements.txt
alembic==1.4.3
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
cffi==1.14.3
click==7.1.2
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-BabelEx==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Principal==0.4.0
Flask-Security-Too==3.4.4
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gevent==20.6.2
greenlet==0.4.16
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
mpmath==1.1.0
passlib==1.7.4
pycparser==2.20
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
six==1.15.0
speaklater==1.3
SQLAlchemy==1.3.19
sympy==1.6.2
Werkzeug==1.0.1
WTForms==2.3.3
zope.event==4.4
zope.interface==5.1.0
